+++ src/shapes/IText/IText.test.ts
@@ -1,3 +1,4 @@
+import { Canvas } from '../../canvas/Canvas';
 import { Group } from '../Group';
 import { IText } from './IText';
 
@@ -41,5 +42,16 @@ describe('IText', () => {
         });
       },
     );
+  });
+  describe('Interaction with mouse and editing', () => {
+    it('_mouseDownHandlerBefore set up selected property', () => {
+      const iText = new IText('test need some word\nsecond line');
+      iText.canvas = new Canvas();
+      expect(iText.selected).toBe(undefined);
+      iText._mouseDownHandler({ e: { button: 0 }, alreadySelected: false });
+      expect(iText.selected).toBe(false);
+      iText._mouseDownHandler({ e: {}, alreadySelected: true });
+      expect(iText.selected).toBe(true);
+    });
   });
 });