+++ src/shapes/IText/ITextClickBehavior.ts
@@ -1,7 +1,9 @@
-import type { TPointerEvent, TPointerEventInfo } from '../../EventTypeDefs';
-import type { XY } from '../../Point';
+import type {
+  ObjectPointerEvents,
+  TPointerEvent,
+  TPointerEventInfo,
+} from '../../EventTypeDefs';
 import { Point } from '../../Point';
-import { stopEvent } from '../../util/dom_event';
 import { invertTransform } from '../../util/misc/matrix';
 import { DraggableTextDelegate } from './DraggableTextDelegate';
 import type { ITextEvents } from './ITextBehavior';
@@ -19,28 +21,14 @@ export abstract class ITextClickBehavior<
   SProps extends SerializedTextProps = SerializedTextProps,
   EventSpec extends ITextEvents = ITextEvents,
 > extends ITextKeyBehavior<Props, SProps, EventSpec> {
-  private declare __lastSelected: boolean;
-  private declare __lastClickTime: number;
-  private declare __lastLastClickTime: number;
-  private declare __lastPointer: XY | Record<string, never>;
-  private declare __newClickTime: number;
-
   protected draggableTextDelegate: DraggableTextDelegate;
 
   initBehavior() {
     // Initializes event handlers related to cursor or selection
     this.on('mousedown', this._mouseDownHandler);
-    this.on('mousedown:before', this._mouseDownHandlerBefore);
     this.on('mouseup', this.mouseUpHandler);
     this.on('mousedblclick', this.doubleClickHandler);
-    this.on('tripleclick', this.tripleClickHandler);
-
-    // Initializes "dbclick" event handler
-    this.__lastClickTime = +new Date();
-    // for triple click
-    this.__lastLastClickTime = +new Date();
-    this.__lastPointer = {};
-    this.on('mousedown', this.onMouseDown);
+    this.on('mousetripleclick', this.tripleClickHandler);
 
     this.draggableTextDelegate = new DraggableTextDelegate(
       this as unknown as IText,
@@ -78,35 +66,6 @@ export abstract class ITextClickBehavior<
   }
 
   /**
-   * Default event handler to simulate triple click
-   * @private
-   */
-  onMouseDown(options: TPointerEventInfo) {
-    if (!this.canvas) {
-      return;
-    }
-    this.__newClickTime = +new Date();
-    const newPointer = options.pointer;
-    if (this.isTripleClick(newPointer)) {
-      this.fire('tripleclick', options);
-      stopEvent(options.e);
-    }
-    this.__lastLastClickTime = this.__lastClickTime;
-    this.__lastClickTime = this.__newClickTime;
-    this.__lastPointer = newPointer;
-    this.__lastSelected = this.selected && !this.getActiveControl();
-  }
-
-  isTripleClick(newPointer: XY) {
-    return (
-      this.__newClickTime - this.__lastClickTime < 500 &&
-      this.__lastClickTime - this.__lastLastClickTime < 500 &&
-      this.__lastPointer.x === newPointer.x &&
-      this.__lastPointer.y === newPointer.y
-    );
-  }
-
-  /**
    * Default handler for double click, select a word
    */
   doubleClickHandler(options: TPointerEventInfo) {
@@ -114,6 +73,7 @@ export abstract class ITextClickBehavior<
       return;
     }
     this.selectWord(this.getSelectionStartFromPointer(options.e));
+    this.renderCursorOrSelection();
   }
 
   /**
@@ -124,6 +84,7 @@ export abstract class ITextClickBehavior<
       return;
     }
     this.selectLine(this.getSelectionStartFromPointer(options.e));
+    this.renderCursorOrSelection();
   }
 
   /**
@@ -134,7 +95,7 @@ export abstract class ITextClickBehavior<
    * initializing a mousedDown on a text area will cancel fabricjs knowledge of
    * current compositionMode. It will be set to false.
    */
-  _mouseDownHandler({ e }: TPointerEventInfo) {
+  _mouseDownHandler({ e, alreadySelected }: ObjectPointerEvents['mousedown']) {
     if (
       !this.canvas ||
       !this.editable ||
@@ -150,7 +111,7 @@ export abstract class ITextClickBehavior<
 
     this.canvas.textEditingManager.register(this);
 
-    if (this.selected) {
+    if (alreadySelected) {
       this.inCompositionMode = false;
       this.setCursorByClick(e);
     }
@@ -162,28 +123,16 @@ export abstract class ITextClickBehavior<
       }
       this.renderCursorOrSelection();
     }
-  }
-
-  /**
-   * Default event handler for the basic functionalities needed on mousedown:before
-   * can be overridden to do something different.
-   * Scope of this implementation is: verify the object is already selected when mousing down
-   */
-  _mouseDownHandlerBefore({ e }: TPointerEventInfo) {
-    if (!this.canvas || !this.editable || notALeftClick(e)) {
-      return;
-    }
-    // we want to avoid that an object that was selected and then becomes unselectable,
-    // may trigger editing mode in some way.
-    this.selected = this === this.canvas._activeObject;
+    this.selected ||= alreadySelected || this.isEditing;
   }
 
   /**
    * standard handler for mouse up, overridable
    * @private
    */
-  mouseUpHandler({ e, transform }: TPointerEventInfo) {
+  mouseUpHandler({ e, transform }: ObjectPointerEvents['mouseup']) {
     const didDrag = this.draggableTextDelegate.end(e);
+
     if (this.canvas) {
       this.canvas.textEditingManager.unregister(this);
 
@@ -195,6 +144,7 @@ export abstract class ITextClickBehavior<
         return;
       }
     }
+
     if (
       !this.editable ||
       (this.group && !this.group.interactive) ||
@@ -205,17 +155,13 @@ export abstract class ITextClickBehavior<
       return;
     }
 
-    if (this.__lastSelected && !this.getActiveControl()) {
-      this.selected = false;
-      this.__lastSelected = false;
+    if (this.selected && !this.getActiveControl()) {
       this.enterEditing(e);
       if (this.selectionStart === this.selectionEnd) {
         this.initDelayedCursor(true);
       } else {
         this.renderCursorOrSelection();
       }
-    } else {
-      this.selected = true;
     }
   }