+++ src/shapes/Image.ts
@@ -781,6 +781,7 @@ export class FabricImage<
     'height',
     'preserveAspectRatio',
     'xlink:href',
+    'href',
     'crossOrigin',
     'image-rendering',
   ];
@@ -851,7 +852,7 @@ export class FabricImage<
       cssRules,
     );
     return this.fromURL(
-      parsedAttributes['xlink:href'],
+      parsedAttributes['xlink:href'] || parsedAttributes['href'],
       options,
       parsedAttributes,
     ).catch((err) => {