+++ test/unit/itext_click_behaviour.js
@@ -118,15 +118,6 @@
       assert.equal(iText.getSelectionStartFromPointer(eventData), 5, 'index');
       assert.equal(iText.getSelectionStartFromPointer({ ...eventData, clientY: 20 }), 5, 'index');
     });
-    QUnit.test('_mouseDownHandlerBefore set up selected property', function (assert) {
-      var iText = new fabric.IText('test need some word\nsecond line');
-      assert.equal(iText.selected, undefined, 'iText has no selected property');
-      canvas.setActiveObject(iText);
-      iText.canvas = canvas;
-      iText._mouseDownHandlerBefore({ e: {} });
-      assert.equal(iText.selected, true, 'iText has selected property');
-      assert.equal(iText.__lastSelected, undefined, 'iText has no __lastSelected property');
-    });
     QUnit.test('mouse down aborts cursor animation', function (assert) {
       var iText = new fabric.IText('test need some word\nsecond line', { canvas });
       assert.ok(typeof iText._animateCursor === 'function', 'method is defined');
@@ -138,17 +129,6 @@
       iText._mouseDownHandler({ e: { target: canvas.upperCanvasEl } });
       assert.equal(animate, 1, 'called from enterEditing');
       assert.equal(aborted, 1, 'called from render');
-    });
-    QUnit.test('_mouseUpHandler set selected as true', function (assert) {
-      var iText = new fabric.IText('test');
-      iText.initDelayedCursor = function () { };
-      iText.renderCursorOrSelection = function () { };
-      assert.equal(iText.selected, undefined, 'iText has no selected property');
-      assert.equal(iText.__lastSelected, undefined, 'iText has no __lastSelected property');
-      canvas.setActiveObject(iText);
-      iText.canvas = canvas;
-      iText.mouseUpHandler({ e: {} });
-      assert.equal(iText.selected, true, 'iText has selected property');
     });
     QUnit.test('_mouseUpHandler on a selected object enter edit', function (assert) {
       var iText = new fabric.IText('test');